requirements.txt
azure-ai-formrecognizer
azure-ai-documentintelligence
openai
python-dotenv
requests
PyPDF2



